+++ src/main/java/vn/com/ecotechgroup/erp/repository/OrderProductRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import vn.com.ecotechgroup.erp.entity.OrderProduct;
 
-public interface OrderProductRepository extends JpaRepository<OrderProduct, Integer>{
+public interface OrderProductRepository extends JpaRepository<OrderProduct, Long >{
 
 }
