+++ src/main/java/vn/com/ecotechgroup/erp/repository/PaymentTypeRepository.java
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Repository;
 import vn.com.ecotechgroup.erp.entity.PaymentType;
 
 @Repository
-public interface PaymentTypeRepository extends JpaRepository<PaymentType, Integer>{
+public interface PaymentTypeRepository extends JpaRepository<PaymentType, Long >{
 	
 	@Query("SELECT p FROM PaymentType p "
 			+ "WHERE p.name LIKE %:searchTerm% "
